cloudproviders/vsphere: Don't filter IPs by MAC address
This commit stops the vSphere cloudprovider [in-tree] from filtering MAC
addresses. This feature was introduced years prior prevening NICs with an
inappropriate organizational unique identifiers (OUI) from have its IP
registered in Kubernetes.

The prior list of OUIs was:
  - 00:50:56
  - 00:0c:29

VMware's current ownership[0] features 4 OUIs namely:
  - 00:50:56
  - 00:0c:29
  - 00:1c:14
  - 00:05:69

This causes the vsphere provider to not register IP addresses for any
nodes with the latter two OUIs in their MAC address. Rather than expand
the list, I've removed validation. Validating against VMware's OUI for
the sake of ensuring the NIC is correct is unnecessary and easy to work
around for anyone who wishes to.

[0]: http://standards-oui.ieee.org/oui.txt