cloudproviders/vsphere: Filter IPs by all VMware OUIs
This commit updates the vSphere cloudprovider [in-tree] to know about all
potential VMware MAC addresses. This feature was introduced years prior
prevening NICs with an inappropriate organizational unique identifiers (OUI) from have its IP
registered in Kubernetes.

The prior list of OUIs was:
  - 00:50:56
  - 00:0c:29

VMware's current ownership[0] features 4 OUIs namely:
  - 00:50:56
  - 00:0c:29
  - 00:1c:14
  - 00:05:69

This causes the vsphere provider to not register IP addresses for any
nodes with the latter two OUIs in their MAC address.

The *best* solution is remove validation. It is completely possible for clusters
to have NICs not assigned a conferment MAC address. This should *not* break Kubernetes.

However, removing this filtering will cause the kubelet to register all
interface IPs. For example, the Docker bridge. A better approach would
be to rely on IPs known by vCenter. Since the in-tree provider will
eventually be deprecated, this will move us forward and we can consider
this design choice for the external provider.